pyproject.toml
[tool.poetry]
name = "ff-flight-data-enrichment-system"
version = "0.1.0"
description = "FlyFlat Flight Data Enrichment System Code Exercise"
authors = [
    "Victor H. Garcia <victorhugogarcia13d@gmail.com>"
]
readme = "README.md"
packages = [
    { include = "ff-flight-data-enrichment-system" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = ">=5.2.3,<6.0.0"
httpx = ">=0.28.1,<0.29.0"
fastapi = ">=0.115.13,<0.116.0"
uvicorn = ">=0.34.3,<0.35.0"
pydantic = ">=2.11.7,<3.0.0"
celery = ">=5.5.3,<6.0.0"
redis = ">=6.2.0,<7.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
backoff = ">=2.2.1,<3.0.0"
djangorestframework = "^3.16.0"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"